feat(ui): remove selectionPasted action
